refactor(package): 데이터를 전달하는 객체를 dto 패키지로 분리
