+++ src/test/java/racingcar/domain/CarsTest.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.vo.Name;
-import racingcar.vo.Winners;
+import dto.Winners;
 import static org.assertj.core.api.Assertions.*;
 
 public class CarsTest {
